refactor: refactor MS Device Code Flow provider structure
