+++ liminal/auth/microsoft/__init__.py
@@ -1,0 +1,1 @@
+"""Define Microsoft providers."""
