+++ liminal/auth/microsoft/models.py
@@ -8,22 +8,6 @@ from typing import Literal
 from liminal.helpers.model import BaseResponseModel
 
 
-class AuthProvider:
-    """Define an auth provider abstract base class."""
-
-    async def get_access_token(self) -> str:
-        """Retrieve an access token from the auth provider.
-
-        The working principle here is that the auth provider will return an access
-        token that can be used to authenticate with the Liminal API server.
-
-        Returns:
-            The access token.
-
-        """
-        raise NotImplementedError
-
-
 @dataclass(frozen=True, kw_only=True)
 class MSALCacheTokenResponse(BaseResponseModel):
     """Define an MSAL token response from Entra ID."""
