+++ tests/auth/test_liminal.py
@@ -14,7 +14,7 @@ import pytest
 from pytest_httpx import HTTPXMock
 
 from liminal import Client
-from liminal.endpoints.auth import MicrosoftAuthProvider
+from liminal.auth.microsoft.device_code_flow import DeviceCodeFlowProvider
 from liminal.errors import AuthError
 from tests.common import TEST_API_SERVER_URL, TEST_CLIENT_ID, TEST_TENANT_ID
 
@@ -72,7 +72,7 @@ async def test_auth_via_refresh_token_new_client(
         ],
     )
 
-    microsoft_auth_provider = MicrosoftAuthProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
+    microsoft_auth_provider = DeviceCodeFlowProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
     async with httpx.AsyncClient() as httpx_client:
         client = Client(
             microsoft_auth_provider, TEST_API_SERVER_URL, httpx_client=httpx_client
@@ -209,7 +209,7 @@ async def test_premature_refresh_token(httpx_mock: HTTPXMock, patch_msal: None) 
         patch_msal: Ensure the MSAL library is patched.
 
     """
-    microsoft_auth_provider = MicrosoftAuthProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
+    microsoft_auth_provider = DeviceCodeFlowProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
     async with httpx.AsyncClient() as httpx_client:
         client = Client(
             microsoft_auth_provider, TEST_API_SERVER_URL, httpx_client=httpx_client
