+++ tests/auth/test_microsoft.py
@@ -9,7 +9,7 @@ import pytest
 from pytest_httpx import HTTPXMock
 
 from liminal import Client
-from liminal.endpoints.auth import MicrosoftAuthProvider
+from liminal.auth.microsoft.device_code_flow import DeviceCodeFlowProvider
 from liminal.errors import AuthError
 from tests.common import TEST_API_SERVER_URL, TEST_CLIENT_ID, TEST_TENANT_ID
 
@@ -35,7 +35,7 @@ async def test_auth_via_device_code_flow(
         patch_msal: Ensure the MSAL library is patched.
 
     """
-    microsoft_auth_provider = MicrosoftAuthProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
+    microsoft_auth_provider = DeviceCodeFlowProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
 
     async with httpx.AsyncClient() as httpx_client:
         client = Client(
@@ -64,7 +64,7 @@ async def test_auth_via_device_code_flow_timeout(
         patch_msal: Ensure the MSAL library is patched.
 
     """
-    microsoft_auth_provider = MicrosoftAuthProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
+    microsoft_auth_provider = DeviceCodeFlowProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
 
     async with httpx.AsyncClient() as httpx_client:
         client = Client(
