+++ tests/conftest.py
@@ -14,7 +14,7 @@ import pytest_asyncio
 from pytest_httpx import HTTPXMock
 
 from liminal import Client
-from liminal.endpoints.auth import MicrosoftAuthProvider
+from liminal.auth.microsoft.device_code_flow import DeviceCodeFlowProvider
 from tests.common import (
     TEST_API_SERVER_URL,
     TEST_CLIENT_ID,
@@ -61,7 +61,9 @@ def _patch_msal_fixture(
         None.
 
     """
-    with patch("liminal.endpoints.auth.PublicClientApplication") as msal_app:
+    with patch(
+        "liminal.auth.microsoft.device_code_flow.PublicClientApplication"
+    ) as msal_app:
         msal_app.return_value.get_accounts = Mock(return_value=msal_accounts)
         msal_app.return_value.initiate_device_flow = Mock(
             return_value={
@@ -120,7 +122,7 @@ async def mock_client_fixture(
         ],
     )
 
-    microsoft_auth_provider = MicrosoftAuthProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
+    microsoft_auth_provider = DeviceCodeFlowProvider(TEST_TENANT_ID, TEST_CLIENT_ID)
     async with httpx.AsyncClient() as httpx_client:
         client = Client(
             microsoft_auth_provider, TEST_API_SERVER_URL, httpx_client=httpx_client
